Our ABI for hostapd currently uses an old structure that only has one byte for the key index.  This means that the 'magic value' of -1 (255) for NONE is passed rather than the larger type we actually use internally.  To make comparisons work properly, I'm putting the magic value back at (uint8_t)-1 again so that a key index of 'none' with a MAC address that is passed into madwifi is interpreted correctly and the keys for that station are properly removed, rather than errors returned.
git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@2798 0192ed92-7a03-0410-a25b-9323aeb14dbd

+++ net80211/ieee80211_crypto.h
@@ -106,7 +106,7 @@ struct ieee80211_key {
 
 #define	IEEE80211_CIPHER_MAX		(IEEE80211_CIPHER_NONE + 1)
 
-#define	IEEE80211_KEYIX_NONE	((ieee80211_keyix_t) -1)
+#define	IEEE80211_KEYIX_NONE	((ieee80211_keyix_t) (uint8_t)-1)
 
 #if defined(__KERNEL__) || defined(_KERNEL)
